requirements.txt
-e git+http://github.com/BlackEarth/bf@cd7c0e7b8b77ebfb8afe4a99b4a35e648be931fa#egg=bf
-e git+https://github.com/BlackEarth/bl.git@1a34b56d80314bd358641d0b882137186c5b09de#egg=bl
cssselect==0.9.2
cssutils==1.0.1
libsass==0.12.3
six==1.10.0
Unum==4.1.3
